Implements a CloudFormation template
TL;DR
-----

Implements a CloudFormation template the manages a Replicated license

Details
-------

Uses Terraform and Python to implement a CloudFormation template that
manages a Replicated customer. This is the foundation for creating an
AWS Marketplace product that is distributed with Replicated. Expanding
this stack with additiona resources like an EC2 instance build from a
Replicated Embedded Cluster AMI will enable selling your product in
the Marketplace.

The license is handled as a custom resource in CloudFormation. It uses
the Replicated customer ID as the resource ID to faciliatte managing the
lifecycle of the customer entirely within CloudFormation. The template
invokes a lambda function to manage the custom resource.

The lambda function is written in Python and use the `crhelper` module
provided by Amazon to clean up the code. Their is a class `Customer`
that manages the customer and uses the Replicated Vendor Portal API  to
create, load, and delete customers. Update is not yet implemented.

The CloudFormation template is stored in an S3 bucket so it's avaialble
for creating new stacks. The S3 bucket and the lamdba function as both
managed with Terraform, along with a role and policy for the stack
execution.